server/quiz/requirements.txt
django
djangorestframework
requests
psycopg2